Fix int precision issus (fixes #179)

+++ src/libserver/trace.cpp
@@ -38,7 +38,7 @@ Trace::TraceHeader (int layer)
   tv.tv_usec -= started.tv_usec;
   tv.tv_sec -= started.tv_sec;
 
-  printf ("Layer %d [%d:%-*s %d.%03d] ", layer, seq, trace_namelen, name.c_str(), tv.tv_sec,tv.tv_usec/1000);
+  printf ("Layer %d [%d:%-*s %u.%03u] ", layer, seq, trace_namelen, name.c_str(), (unsigned int)tv.tv_sec,(unsigned int)tv.tv_usec/1000);
 }
 
 void
